refactor(tauri-plugin): move the five inline test modules to sibling files (#1918)
Ninth and last crate of the campaign. All five blocks are a plain
#[cfg(test)] mod tests at end of file, top-level, with no sibling to
collide with, so every one is a straight #[path] move.

Verification here is weaker than for the other eight crates, and the
difference is worth stating rather than glossing: this crate does not
build in this environment — cargo check fails in the gdk-sys build
script for want of the GTK development libraries — so there is no
test-name list to diff, no cargo test and no clippy. What did run:
rustfmt (which only parses) is clean, and the three text-level guards
pass. The compile-level proof rests on CI.

With this the inline-tests baseline reaches zero: 173 files and 21788
frozen lines at the start of #1918, none left.
